refactor: substrate network form
