+++ apps/extension/src/ui/apps/dashboard/routes/Networks/NetworkPage.tsx
@@ -1,5 +1,9 @@
 import { AnalyticsPage } from "@ui/api/analytics"
-import { EvmNetworkForm, SubNetworkForm } from "@ui/domains/Settings/ManageNetworks/NetworkForm"
+import {
+  EvmNetworkForm,
+  SubNetworkFormAdd,
+  SubNetworkFormEdit,
+} from "@ui/domains/Settings/ManageNetworks/NetworkForm"
 import { useAnalyticsPageView } from "@ui/hooks/useAnalyticsPageView"
 import { useCallback } from "react"
 import { useTranslation } from "react-i18next"
@@ -38,7 +42,12 @@ export const NetworkPage = () => {
 
   return (
     <DashboardLayout analytics={ANALYTICS_PAGE} withBack centered>
-      {isChain && <SubNetworkForm chainId={id} onSubmitted={handleSubmitted} />}
+      {isChain && (
+        <>
+          {id && <SubNetworkFormEdit chainId={id} onSubmitted={handleSubmitted} />}
+          {!id && <SubNetworkFormAdd onSubmitted={handleSubmitted} />}
+        </>
+      )}
       {isEvmNetwork && <EvmNetworkForm evmNetworkId={id} onSubmitted={handleSubmitted} />}
     </DashboardLayout>
   )
