+++ apps/extension/src/ui/domains/Settings/ManageNetworks/NetworkForm/Substrate/RemoveEvmNetworkButton.tsx
