+++ apps/extension/src/ui/domains/Settings/ManageNetworks/NetworkForm/Substrate/index.ts
@@ -1,0 +1,2 @@
+export { SubNetworkFormAdd } from "./Add"
+export { SubNetworkFormEdit } from "./Edit"
