+++ apps/extension/src/ui/domains/Settings/ManageNetworks/NetworkForm/Substrate/types.ts
@@ -1,0 +1,3 @@
+export type SubNetworkFormBaseProps = {
+  onSubmitted?: () => void
+}
