+++ apps/extension/src/ui/domains/Settings/ManageNetworks/NetworkForm/index.ts
@@ -1,2 +1,2 @@
-export { SubNetworkForm } from "./SubNetworkForm"
+export { SubNetworkFormEdit, SubNetworkFormAdd } from "./Substrate"
 export { EvmNetworkForm } from "./EvmNetworkForm"
