+++ apps/extension/src/ui/domains/Settings/ManageNetworks/NetworkForm/useRegisterFieldWithDebouncedValidation.ts
@@ -1,6 +1,7 @@
 import debounce from "lodash/debounce"
 import { useCallback } from "react"
 import {
+  ChangeHandler,
   FieldPath,
   FieldValues,
   RegisterOptions,
@@ -10,24 +11,35 @@ import {
 } from "react-hook-form"
 
 // inspired from https://github.com/react-hook-form/react-hook-form/issues/40
-export const useRegisterFieldWithDebouncedValidation = <TFieldValues extends FieldValues>(
-  name: FieldPath<TFieldValues>,
+export const useRegisterFieldWithDebouncedValidation = <
+  TFieldValues extends FieldValues,
+  TFieldPath extends FieldPath<TFieldValues>
+>(
+  name: TFieldPath,
   delay: number,
   trigger: UseFormTrigger<TFieldValues>,
   register: UseFormRegister<TFieldValues>,
+  extraValidationCb?: (name: TFieldPath, value: FieldValues[TFieldPath]) => Promise<void>,
   options?: RegisterOptions<TFieldValues, FieldPath<TFieldValues>>
 ) => {
   const useFormRegisterReturn: UseFormRegisterReturn = register(name, options)
   const { onChange } = useFormRegisterReturn
 
   // eslint-disable-next-line react-hooks/exhaustive-deps
-  const handleChange = useCallback(
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    debounce((e: any) => {
-      onChange(e)
+  const debouncedTrigger = useCallback(
+    debounce(async (value: FieldValues[TFieldPath]) => {
       trigger(name)
+      extraValidationCb && (await extraValidationCb(name, value))
     }, delay),
-    [name, trigger, onChange, delay]
+    [trigger, extraValidationCb, delay]
+  )
+
+  const handleChange = useCallback(
+    (e: Parameters<ChangeHandler>[0]) => {
+      onChange(e)
+      debouncedTrigger(e.target.value)
+    },
+    [onChange, debouncedTrigger]
   )
 
   return {
